// lib/types/UserTypes.ts
export type UserDetailsT = {
  name: string;
  phone: string;
  email: string;
  gender: string;
  nation: string;
  state: number;
  town: string;
  citizenship: string;
  nrc_number: number;
  payment: string;
  pickup_place: string;
  total_seat: number;
  selected_seat_no: string[];
  ticket_id: number;
  id: number;
};